move dashboard to its own directory, split more components
